requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets>=12.0
watchfiles>=1.0.0
requests==2.31.0
rasterio==1.3.9
geopandas==0.14.1
numpy==1.24.3
shapely==2.0.1
python-dotenv==1.0.0
pydantic==2.5.0
Pillow==10.1.0
sqlalchemy==2.0.23
psycopg[binary]==3.2.10
earthengine-api>=0.1.384
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
mutagen>=1.47.0
firebase-admin>=6.5.0
